requirements.txt
Django==2.2
whitenoise
flake8
Pillow
black
coveralls
mysqlclient
six
PyYAML
python-dateutil
